+++ test/environments/distributions/runtests.jl
@@ -327,4 +327,5 @@ end
 end
 
 include("distribution_expansion.jl")
+include("distribution_transforms.jl")
 include("categorical.jl")